agent-dir/agents/compliance-agent/requirements.txt
# compliance-agent Dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
numpy>=1.24.0
# Base agent dependencies
-e ./base_agent

# Compliance-specific dependencies
langchain>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
Flask==2.3.3
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
selenium>=4.15.0
urllib3>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake-useragent>=1.4.0